requirements.txt
asgiref==3.9.2
Django==5.2.6
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.16.1
MarkupSafe==3.0.2
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
Werkzeug==3.1.3
